# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forafix API Documentation",
        "description": "Marketplace API with Wallet, Escrow and Booking management",
        "contact": {
            "email": "support@forafix.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Send password reset link to email",
                "operationId": "b7005af67cef64028a6bbf0c4256f136",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Reset password using token",
                "operationId": "b6e7b32be0980ee2d6c1440030814304",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful"
                    }
                }
            }
        },
        "/api/location/reverse-geocode": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Convert coordinates to address/area name",
                "operationId": "5f5eec5055d5bf9d4e36bf00e8c623dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.0765
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 7.3986
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/location/search": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Search for places/addresses (Google Autocomplete)",
                "operationId": "8e6dd66a6fe9b789634db367dc216e4e",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/location/place-details/{placeId}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get specific place details by ID",
                "operationId": "7a196cb433d8d2fd242cbdcbed24be6e",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}